+++ docs/changelog.rst
@@ -9,6 +9,7 @@ CHANGELOG
 
 * Use POI pictures in PDF if the trek has no picture itself
 * Use a placeholder in PDF if there is no picture
+* Parser to import touristic contents from SITRA
 
 
 2.7.2 (2016-01-26)